pyproject.toml
[tool.poetry]
name = "rag-chatbot"
version = "0.1.0"
description = ""
authors = ["Yash Deshpande <4883988-YashDeshpande1@users.noreply.replit.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
flask = "3.1.0"
# openai = "1.75.0"  # Commented out OpenAI dependency
langchain = "0.3.24"
# langchain-openai = "0.3.14"  # Commented out OpenAI LangChain dependency
langchain-google-genai = "2.1.8"  # Added Google Gemini LangChain dependency
langchain-community = "0.3.22"
langchain-text-splitters = "0.3.8"
gunicorn = "23.0.0"
pypdf = "5.4.0"
pdfminer-six = "20250416"
python-docx = "1.1.2"
unstructured = "0.12.4"
trafilatura = "2.0.0"
python-multipart = "0.0.20"
werkzeug = "3.1.3"
openpyxl = "3.1.5"
email-validator = "2.1.1"
python-dotenv = "^1.1.0"
pinecone-client = "^3.0.0"  # Added Pinecone for persistent storage

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
